moving data to qualifierFormData
