+++ services/ui-src/src/measures/2024/shared/Qualifiers/deliverySystems.tsx
@@ -7,7 +7,7 @@ import { BsPercent } from "react-icons/bs";
 import { percentageAllowOneDecimalMax } from "utils";
 import { useUser } from "hooks/authHooks";
 import { UserRoles } from "types";
-import { DataDriven } from "./data";
+import { DataDriven } from "labels/2024/qualifierFormsData";
 
 const initialDeliverySystemValue = {
   label: "",
